requirements.txt
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
